fast-tax-python/requirements.txt
certifi==2025.7.9
charset-normalizer==3.4.2
idna==3.10
pip==25.1.1
requests==2.32.4
setuptools==49.2.1
suds-community==1.2.0
urllib3==2.5.0
